+++ src/main/java/com/example/demo/Services/AppointmentService.java
@@ -1,8 +1,10 @@
 package com.example.demo.Services;
 
-import com.example.demo.DTOs.BookAppointmentRequest;
-import com.example.demo.DTOs.AppointmentResponse;
-import com.example.demo.Entities.AppointmentStatus;
+
+import com.example.demo.DTOs.RequestDTO.BookAppointmentRequest;
+import com.example.demo.DTOs.ResponseDTO.AppointmentResponse;
+import com.example.demo.Enums.AppointmentStatus;
+
 import java.util.List;
 
 public interface AppointmentService {
@@ -15,20 +17,20 @@ public interface AppointmentService {
     /**
      * Cập nhật trạng thái vòng đời của một lịch hẹn (PENDING, CONFIRMED, COMPLETED...)
      */
-    AppointmentResponse updateStatus(Long appointmentId, AppointmentStatus status);
+    AppointmentResponse updateStatus(Integer appointmentId, AppointmentStatus status);
 
     /**
      * Hủy lịch hẹn và giải phóng slot khám trống cho bác sĩ
      */
-    AppointmentResponse cancelAppointment(Long appointmentId, String reason);
+    AppointmentResponse cancelAppointment(Integer appointmentId, String reason);
 
     /**
      * Lấy danh sách lịch sử khám bệnh của một Bệnh nhân (Sắp xếp theo ngày mới nhất)
      */
-    List<AppointmentResponse> getPatientAppointments(Long patientId);
+    List<AppointmentResponse> getPatientAppointments(Integer patientId);
 
     /**
      * Lấy danh sách lịch hẹn cần khám của một Bác sĩ
      */
-    List<AppointmentResponse> getDoctorAppointments(Long doctorId);
+    List<AppointmentResponse> getDoctorAppointments(Integer doctorId);
 }
